+++ x/sp/keeper/sp.go
@@ -124,6 +124,7 @@ func (k Keeper) GetAllStorageProviders(ctx sdk.Context) (sps []types.StorageProv
 	store := ctx.KVStore(k.storeKey)
 
 	iter := storetypes.KVStorePrefixIterator(store, types.StorageProviderKey)
+	defer iter.Close()
 
 	for ; iter.Valid(); iter.Next() {
 		sp := types.MustUnmarshalStorageProvider(k.cdc, iter.Value())
@@ -140,7 +141,7 @@ func (k Keeper) Exit(ctx sdk.Context, sp *types.StorageProvider) error {
 	store.Delete(types.GetStorageProviderByApprovalAddrKey(sdk.MustAccAddressFromHex(sp.ApprovalAddress)))
 	store.Delete(types.GetStorageProviderByGcAddrKey(sdk.MustAccAddressFromHex(sp.GcAddress)))
 	store.Delete(types.GetStorageProviderKey(k.spSequence.EncodeSequence(sp.Id)))
-	store.Delete(types.GetStorageProviderByBlsKeyKey(types.GetStorageProviderByBlsKeyKey(sp.GetBlsKey())))
+	store.Delete(types.GetStorageProviderByBlsKeyKey(sp.GetBlsKey()))
 	return nil
 }